Split Solid::Process::EventLogsRecord into multiple files
